+++ sphinx-docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, RAISE Lab, NC State University'
 author = 'RAISE Lab, NC State University'
 
 # The full version, including alpha/beta/rc tags
-release = '1.6.3'
+release = '2.0.0'
 
 
 # -- General configuration ---------------------------------------------------